+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - 修改 XEnv、XLog 模块对新版 XPrefs.IPanel 接口的适配
 - 修改 XComp 若干函数的名称：CloneGO -> CloneGameObject、DestroyGO -> DestroyGameObject、SetActiveState -> SetGameObjectActive
 - 修改 XLoom 若干函数的名称：StartCR -> StartCoroutine、StopCR -> StopCoroutine
+- 修改 XString 的 IEval 接口的名称为 IEvaluator
 
 ## [0.1.7] - 2025-07-07
 ### 变更